Fix more consistent error handling in project API settings page
