--- studio/components/ui/ProjectSettings/SettingsLoadingState.types.tsx
@@ -1,4 +1,0 @@
-export interface SettingsLoadingStateProps {
-  isError: boolean
-  errorMessage: string
-}
