+++ studio/hooks/queries/index.ts
@@ -8,3 +8,4 @@ export * from './useProjectPaygStatistics'
 export * from './useProjectSettings'
 export * from './useProjectSubscription'
 export * from './useSubscriptionStats'
+export * from './useProjectPostgrestConfig'
